Update mmock to a newer version
Go get tries to get a version of mmock that is not compatible with
Gandalf before this change due to some refactoring that was done to its
internals. For some reason these definitions marshal to values that they
cannot be unmarshaled from, this fixes that so we have to mangle them a
little.